Use shell pwd instead of PWD in Makefile
When doing make -C subdir, the value of the PWD environment
variable is inherited from the shell process and remains unchanged when
make enters subdir. This will cause the various path variables to be
incorrect and cause the build to fail.

If we use $(shell pwd) instead, the working directory will be as
expected. That is, it will point to the subdirectory.