Only recompile Args.cpp when needed
Previously, Args.cpp was always recompiled. This was done to ensure that
the current git SHA was compiled into the source even if the object file
appeared to be up-to-date according to the timestamps.

The down side of this approach is that it means running make will always
take several seconds to compile Args.cpp and link it with the entire
program even if no changes have been made.

This change will recompile Args.cpp only when necessary. That is, it
will recompile Args.cpp if and only if either the object file is
out-of-date or the git SHA in the object file doesn't match the current
git SHA.