Change how to check for presence of IMSIC
When an APLIC domain is configured for direct delivery, we only check
the eidelivery in IMSIC if an IMSIC is present.

This had been done by checking the aiaEnabled flag in CsRegs, but this
actually doesn't work because aiaEnabled will be true even if there is
only an APLIC.

Instead, we can check the "imsic_" pointer in each hart. If the config
json doesn't specify an IMSIC configuration, these pointers will be
null.