Use _GLIBCXX_USE_CXX11_ABI=0 to link with boost
The boost libraries at /tools_vendor/FOSS/boost/1.82 were compiled using
an old ABI. If gcc uses _GLIBCXX_USE_CXX11_ABI=1 while compiling
whisper, it will fail to link against this library.

On the login machines running CentOS, 0 is the default value for this
macro, but on the RHEL8 machines (e.g. ausc-rv-l-13, etc.), gcc uses 1
by default.

Thus, in order to get whisper to compile and link successfully on these
machines, we must either rebuild boost with the newer ABI (preferable),
or instruct gcc to use the older ABI while compiling whisper.

There does not appear to be a good way to determine what ABI a library
uses (you can grep the output of nm -C and look for
std::__cxx11::basic_string but I'd rather not put that in the Makefiles)
so for now this commit checks if the BOOST_ROOT points to this specific
path and if so uses the old ABI.

Perhaps in the future we can rebuild boost with the newer ABI and remove
this altogether.