Make sourcecfg writes conform to AIA section 4.7
Previous to this change, writes to sourcecfg would not clear the pending
bit for level-sensitive interrupts even if the write caused the
rectified input value of the source to become low under the new source
mode. This was based on a misinterpretation of this sentence from the
AIA spec:

> A write to a sourcecfg register will not by itself cause a pending bit
> to be cleared except when the source is made inactive.

With this change, writes to sourcecfg follow the rules in section 4.7.
Specifically, for level-sensitive interrupts, the pending bit is cleared
if the rectified input value is low.